Attempted to run chess engines on the Teensy.
I kept running into memory issues when compiling the code for the ESP8266, but limitations in the stack. Micromax4.8 (mm4.8) was able to run on the Teensy4.1 with minor change to print the bard over Serial, but there is something seriously wrong with the move evaluation, because no moves are accepted for the black side, and moves with Knights are weighted super heavily.